git-svn-id: https://svn.apache.org/repos/asf/accumulo/branches/ACCUMULO-259@1404662 13f79535-47bb-0310-9956-ffa450edef68

+++ server/src/main/java/org/apache/accumulo/server/test/randomwalk/multitable/Commit.java
@@ -27,8 +27,8 @@ public class Commit extends Test {
   public void visit(State state, Properties props) throws Exception {
     state.getMultiTableBatchWriter().flush();
     
-    Integer numWrites = state.getInteger("numWrites");
-    Integer totalWrites = state.getInteger("totalWrites") + numWrites;
+    Long numWrites = state.getLong("numWrites");
+    Long totalWrites = state.getLong("totalWrites") + numWrites;
     
     log.debug("Committed " + numWrites + " writes.  Total writes: " + totalWrites);
